Refactored with simplifications and options
Switched to built-ins for executing EXEs and writing the concatenation file. Compose all paths with the built-in path library for consistency with separators. Added a switch for different output modes based on input arguments (bounce, shutter, sporadic). Temp files and folders are now contained within one parent folder for cleanliness and ease of removal with recursion. Output writes with the original filename plus the resize type to the same folder location. All non-dynamic path locations are composed in one place for ease of examination. Reduced to one call to ffprobe, which returns JSON that is deconstructed. Initial setup for expecting audio-less input. Sorting is frames only breaks down the filename once per file instead of multiple times through the sorting process.

There's definitely room for a lot more improvement. I would have liked to have another argument for the bounce rate, though without a library it might have gotten a bit too complicated. I also believe it's possible to pull the audio directly from the source instead of having an intermediate audio file, which would reduce the ffmpeg calls further, though I wasn't sure the correct arguments. Also the output filename should probably be sanitized a bit for length just in case, though it varies by platform. The removal of the semicolons and switch from apostrophes to quotation marks is a personal preference.