Uses the ytdl-core library to download a YouTube video directly and then converts it to its respective webm.
First validates whether the URL provided is a valid YouTube video URL and then downloads it into a temporary download folder created within the main temporary workLocations folder. It then uses this downloaded video as the source video to convert to a webm. Everything else within the script is left untouched.